fix account management flows naming
