--- fearless/Modules/AccountChange/ChangeMainWireframe.swift
@@ -1,32 +1,0 @@
-import Foundation
-
-final class ChangeMainWireframe: OnboardingMainWireframeProtocol {
-    func showSignup(from view: OnboardingMainViewProtocol?) {
-        guard let usernameSetup = UsernameSetupViewFactory.createViewForSwitch() else {
-            return
-        }
-
-        if let navigationController = view?.controller.navigationController {
-            navigationController.pushViewController(usernameSetup.controller, animated: true)
-        }
-    }
-
-    func showAccountRestore(from view: OnboardingMainViewProtocol?) {
-        guard let restorationController = AccountImportViewFactory.createViewForSwitch()?.controller else {
-            return
-        }
-
-        if let navigationController = view?.controller.navigationController {
-            navigationController.pushViewController(restorationController, animated: true)
-        }
-    }
-
-    func showKeystoreImport(from view: OnboardingMainViewProtocol?) {
-        if
-            let navigationController = view?.controller.navigationController,
-            navigationController.topViewController == view?.controller,
-            navigationController.presentedViewController == nil {
-            showAccountRestore(from: view)
-        }
-    }
-}
