--- fearless/Modules/AccountChange/ChangeUsernameWireframe.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-final class ChangeUsernameWireframe: UsernameSetupWireframeProtocol {
-    func proceed(from view: UsernameSetupViewProtocol?, username: String) {
-        guard let accountCreation = AccountCreateViewFactory.createViewForSwitch(username: username) else {
-            return
-        }
-
-        view?.controller.navigationController?.pushViewController(accountCreation.controller,
-                                                                  animated: true)
-    }
-}
