+++ fearless/Modules/AddAccount/AddAccount.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+struct AddAccount {}
