+++ fearless/Modules/AddAccount/Wireframes/AddAccount+AccountConfirmWireframe.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+extension AddAccount {
+    final class AccountConfirmWireframe: AccountConfirmWireframeProtocol {
+        func proceed(from view: AccountConfirmViewProtocol?) {
+            guard let navigationController = view?.controller.navigationController else {
+                return
+            }
+
+            MainTransitionHelper.transitToMainTabBarController(closing: navigationController,
+                                                               animated: true)
+        }
+    }
+}
