+++ fearless/Modules/AddAccount/Wireframes/AddAccount+UsernameSetupWireframe.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+extension AddAccount {
+    final class UsernameSetupWireframe: UsernameSetupWireframeProtocol {
+        func proceed(from view: UsernameSetupViewProtocol?, username: String) {
+            guard let accountCreation = AccountCreateViewFactory.createViewForAdding(username: username) else {
+                return
+            }
+
+            view?.controller.navigationController?.pushViewController(accountCreation.controller,
+                                                                      animated: true)
+        }
+    }
+
+}
