--- fearless/Modules/AddAccount/Wireframes/AddConfirmationWireframe.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-final class AddConfirmationWireframe: AccountConfirmWireframeProtocol {
-    func proceed(from view: AccountConfirmViewProtocol?) {
-        guard let navigationController = view?.controller.navigationController else {
-            return
-        }
-
-        MainTransitionHelper.transitToMainTabBarController(closing: navigationController,
-                                                           animated: true)
-    }
-}
