+++ fearless/Modules/SelectConnection/SelectConnection.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+struct SelectConnection {}
