+++ fearless/Modules/SwitchAccount/SwitchAccount.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+struct SwitchAccount {}
